backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
# psycopg2-binary  # Only needed for PostgreSQL (install when deploying to prod)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
